Serial: Wait for TX FIFO empty to synchronise UART in flush
Extend the pending operation for the Xilinx XUARTPS to wait on the TX
FIFO to become empty to synchronise the UART / serial communication
between U-Boot and the booted binaries. The pending operation is used by
the serial_flush function called via flush in the boot.c do_go function.
Note, that the U-Boot configuration has to enable the DM_SERIAL and
CONFIG_CONSOLE_FLUSH_SUPPORT as indicated by serial.h (should be enabled
by default for the Xilinx MPSoC ZCU104 evaluation board).